Changing score calculation to be based on individual pair scores
Previously pair combinations were scored by:
- sorting all possible pairs by some criteria
- taking the index of each pair in that list and adding the indices of each pair that appeared in that pair combination
- this gave the final score of a pair combination

This had the bug where pair combinations that were actually equal, based on their component pairs having the same weight, ending up having different scores because of the indices of the list the sorted pairs were stored in.

Now, the different criteria that the pairs were compared with have been translated into weights depending on the criteria and then totalled to give each pair an overall score.

This results in more precision on the scores of a pair combination and the ability to easily identify the exact breakdown of why a combination has that score.

Also included in this change is the removal of sorting the pairs by the names of who is in the pair. This was never done in real life when picking pairs manually.