Making the order of PairStreamEntryPoint#computeScoredCombinations() deterministic
Previously, if two ScoredCombinations had the same total score, when being sorted they were then treated equally. Since the underlying data structure was unordered, this made the resulting order non-deterministic.

Now when sorting we consider the hashCode of the Combination. This has been reimplemented to consider the order in which its underlying components are composed.

See CombinationTest for more information.